add helper.js
